Refactor styled button
