+++ tv-shows/src/components/features/review/ReviewItem/ReviewItem.tsx
@@ -2,43 +2,33 @@
 
 import { Avatar, Card, CardBody, CardFooter, CardHeader, Flex, Text } from '@chakra-ui/react';
 import { IReview } from '../../../../typings/Review.type';
-import { StyledButton } from '../../../core/Button/Button';
+import { StyledButton } from '../../../core/StyledButton/StyledButton';
 import { StarsRating } from '../../../shared/StarsRating/StarsRating';
 
-interface IReviewItemProps extends IReview {
-	onRemoveReview: (review: IReview) => void;
+interface IReviewItemProps {
+	review: IReview;
+	onRemoveReview: (uuid: string) => void;
 }
 
-export const ReviewItem = ({
-	reviewerAvatarURL,
-	reviewerEmail,
-	comment,
-	rating,
-	uuid,
-	onRemoveReview,
-}: IReviewItemProps) => {
+export const ReviewItem = ({ review, onRemoveReview }: IReviewItemProps) => {
 	return (
 		<Card bgColor="purple.700" color="white" borderRadius="2xl" overflow="hidden" mb={4}>
 			<CardHeader>
 				<Flex gap={2} align="center">
-					<Avatar src={reviewerAvatarURL} />
+					<Avatar src={review.reviewerAvatarURL} />
 					<Text data-testid="review-email" fontSize="sm">
-						{reviewerEmail}
+						{review.reviewerEmail}
 					</Text>
 				</Flex>
 			</CardHeader>
 			<CardBody py="0">
 				<Text data-testid="review-comment" fontSize="sm">
-					{comment ?? ' User did not leave a comment.'}
+					{review.comment ?? ' User did not leave a comment.'}
 				</Text>
-				<StarsRating rating={rating} canInteract={false} />
+				<StarsRating rating={review.rating} canInteract={false} />
 			</CardBody>
 			<CardFooter py={0}>
-				<StyledButton
-					label="Remove"
-					review={{ reviewerAvatarURL, reviewerEmail, comment, rating, uuid }}
-					onRemoveReview={onRemoveReview}
-				/>
+				<StyledButton onClick={() => onRemoveReview(review.uuid)}>Remove</StyledButton>
 			</CardFooter>
 		</Card>
 	);
