+++ tv-shows/src/components/features/review/ReviewList/ReviewList.tsx
@@ -1,19 +1,19 @@
-"use client";
+'use client';
 
-import { IReview } from "../../../../typings/Review.type";
-import { ReviewItem } from "../ReviewItem/ReviewItem";
+import { IReview } from '../../../../typings/Review.type';
+import { ReviewItem } from '../ReviewItem/ReviewItem';
 
 interface IReviewListProps {
-  reviews?: Array<IReview>;
-  onRemoveReview: (review: IReview) => void;
+	reviews?: Array<IReview>;
+	onRemoveReview: (uuid: string) => void;
 }
 
 export const ReviewList = ({ reviews, onRemoveReview }: IReviewListProps) => {
-  return (
-    <>
-      {reviews?.map((review, index) => (
-        <ReviewItem key={index} {...review} onRemoveReview={onRemoveReview} />
-      ))}
-    </>
-  );
+	return (
+		<>
+			{reviews?.map((review, index) => (
+				<ReviewItem key={index} review={review} onRemoveReview={onRemoveReview} />
+			))}
+		</>
+	);
 };
